🚨 CRITICAL FIXES: Channel restriction + Admin consolidation + Better error logging
🔧 Critical Fixes:
- Restrict tip detection to #fundraising channel only (ID: 1341173038514704465)
- Add detailed error logging for donation processing failures
- Add detailed error logging for action button failures
- Consolidated admin command to interactive menu system

🎯 Admin Command Overhaul:
- Convert 12 subcommands to 6 interactive categories
- Setup, Dashboard, Analytics, Draw Management, User Management, Features
- Professional pagination with comprehensive data views
- Much cleaner and organized admin interface

🛠️ Error Handling:
- Better error logging with stack traces
- Detailed action button error reporting
- Improved debugging information

Ready for testing channel restriction and admin interface!
